Moved SendMailAttempts model class out of model.py
