+++ iot_api/user_api/models/__init__.py
@@ -23,3 +23,4 @@ from iot_api.user_api.models.GlobalData import GlobalData
 from iot_api.user_api.models.LoginAttempts import LoginAttempts
 from iot_api.user_api.models.PasswordReset import PasswordReset
 from iot_api.user_api.models.RevokedTokenModel import RevokedTokenModel
+from iot_api.user_api.models.SendMailAttempts import SendMailAttempts
